Staff Version
The program should prompt for credits at the pass, defer and fail and display the appropriate progression for each individual student until the staff member user enters ‘q’ to quit. When ‘q’ is entered, the program should produce a ‘histogram’ where each star represents a student who achieved a progress outcome in the category range: progress, trailing, module retriever, and exclude. And finally, display the number of students for each progression category and the total number of outcomes processed.